workers/worker-monitor/requirements.txt
python-telegram-bot==20.7
docker
schedule
